+++ clients/derivatives_trading_portfolio_margin/src/binance_sdk_derivatives_trading_portfolio_margin/rest_api/models/modify_um_order_response.py
@@ -46,9 +46,6 @@ class ModifyUmOrderResponse(BaseModel):
         alias="modifyId",
     )
     price: Optional[StrictStr] = Field(default=None, description="Price.")
-    avg_price: Optional[StrictStr] = Field(
-        default=None, description="Avg Price.", alias="avgPrice"
-    )
     orig_qty: Optional[StrictStr] = Field(
         default=None, description="Orig Qty.", alias="origQty"
     )
@@ -57,9 +54,6 @@ class ModifyUmOrderResponse(BaseModel):
     )
     cum_qty: Optional[StrictStr] = Field(
         default=None, description="Cum Qty.", alias="cumQty"
-    )
-    cum_quote: Optional[StrictStr] = Field(
-        default=None, description="Cum Quote.", alias="cumQuote"
     )
     time_in_force: Optional[StrictStr] = Field(
         default=None, description="Time In Force.", alias="timeInForce"
@@ -103,11 +97,9 @@ class ModifyUmOrderResponse(BaseModel):
         "clientOrderId",
         "modifyId",
         "price",
-        "avgPrice",
         "origQty",
         "executedQty",
         "cumQty",
-        "cumQuote",
         "timeInForce",
         "type",
         "reduceOnly",
@@ -190,11 +182,9 @@ class ModifyUmOrderResponse(BaseModel):
                 "clientOrderId": obj.get("clientOrderId"),
                 "modifyId": obj.get("modifyId"),
                 "price": obj.get("price"),
-                "avgPrice": obj.get("avgPrice"),
                 "origQty": obj.get("origQty"),
                 "executedQty": obj.get("executedQty"),
                 "cumQty": obj.get("cumQty"),
-                "cumQuote": obj.get("cumQuote"),
                 "timeInForce": obj.get("timeInForce"),
                 "type": obj.get("type"),
                 "reduceOnly": obj.get("reduceOnly"),